code/TA4Health_AzureFunc/requirements.txt
azure-functions==1.0.5
pandas==0.25.3
requests==2.22.0